refactoring for streamlit and google cloud
