+++ dataSources/ms2soft.py
@@ -163,7 +163,8 @@ def load_download_records(stationID):
     else:
         dataFolder = path + '/data'
 
-    dataFile = f'{dataFolder}/{stationID}.pkl'
+    dataFile = f'{dataFolder}/{stationID}-raw.pkl'
+    completeFile = f'{dataFolder}/{stationID}-complete.pkl'
     logFile = f'{dataFolder}/{stationID}-log.pkl'
 
     # Load count data
@@ -172,6 +173,14 @@ def load_download_records(stationID):
     except Exception:
         logger.info('Could not find %s, create new dataframe', dataFile)
         stationData = pd.DataFrame()
+
+    
+    # Load cleaned data
+    try:
+        stationComplete = pd.read_pickle(completeFile)
+    except Exception:
+        logger.info('Could not find %s, create new dataframe', completeFile)
+        stationComplete = None
 
     # Load scraped days log
     try:
@@ -199,7 +208,7 @@ def load_download_records(stationID):
     # downloadedDates.drop(downloadedDates.tail(7).index, inplace=True)
     # downloadedDates = downloadedDates.to_list()
 
-    return stationData, dataLog, downloadedDates
+    return stationData, dataLog, downloadedDates, stationComplete
 
 # %% Data Cleaning
 def clean_iframe(iframe_content, stationInfo, scrapeLog):
@@ -293,6 +302,34 @@ def clean_iframe(iframe_content, stationInfo, scrapeLog):
 
     return dfData, scrapeLog
 
+def standardize_df(newdf, fulldf=None):
+    '''Clean up scraped data into consistent format for easier working
+    '''
+    NMDSbike = newdf.xs('Bike', axis=1, level=1).droplevel(level=0, axis=1)
+    NMDSbike['Date'] = newdf.droplevel(level=[1, 2], axis=1)['Date']
+    NMDSbike['Time'] = newdf.droplevel(level=[0, 1], axis=1)['Time']
+
+    NMDSbike['DateTime'] = pd.to_datetime(NMDSbike['Date']+NMDSbike['Time'],
+                                        format=r'%m/%d/%Y%H:%M:%S')
+
+    NMDSbike['Year'] = NMDSbike['DateTime'].dt.year
+    NMDSbike['Month'] = NMDSbike['DateTime'].dt.month
+    NMDSbike['Day'] = NMDSbike['DateTime'].dt.day
+
+    NMDSbike['MonthName'] = NMDSbike['DateTime'].dt.month_name()
+    NMDSbike['DayofWeek'] = NMDSbike['DateTime'].dt.day_name()
+    NMDSbike['MonthApprev'] = NMDSbike['Date'].dt.strftime('%b')
+
+    if fulldf is not None:
+        fulldf = pd.concat([fulldf, NMDSbike])
+    else:
+        fulldf = NMDSbike
+
+    return fulldf
+
+
+# %% Tweeting
+
 def format_tweet(stationID, countData):
     stationName = stations_info[stationID]['TweetName']
 
@@ -319,7 +356,7 @@ def download_all_data(session, stationList):
         scrapeDate = station[1]
 
         # Import pickle of previous downloaded dates and data
-        stationData, stationDataLog, downloadedDates = load_download_records(stationID)
+        stationData, stationDataLog, downloadedDates, stationComplete = load_download_records(stationID)
 
         logger.info('%s: Start downloading %s', datetime.now(), stationID)
 
@@ -355,6 +392,8 @@ def download_all_data(session, stationList):
                         tweetList.append(newTweet)
                     except Exception as e:
                         logger.info('Failed to make tweet: %s', e)
+                    
+                    stationComplete = standardize_df(newData, stationComplete)
 
                 # Add scrape log for date to station scraping log
                 stationDataLog.loc[len(stationDataLog)] = scrapeLog
@@ -363,7 +402,8 @@ def download_all_data(session, stationList):
                 logger.debug('No need to download %s', scrapeDate)
                 scrapeDate += timedelta(days=1)
 
-        stationData.to_pickle(f'data/{stationID}.pkl', protocol=3)
+        stationData.to_pickle(f'data/{stationID}-raw.pkl', protocol=3)
+        stationComplete.to_pickle(f'data/{stationID}-complete.pkl', protocol=3)
         stationDataLog.to_pickle(f'data/{stationID}-log.pkl', protocol=3)
 
         # Save in human readble form for debug
