+++ requirements.txt
@@ -11,7 +11,9 @@ matplotlib==3.6.1
 scipy==1.9.2
 beautifulsoup4==4.10.0
 lxml==4.9.1
-pydrive2==1.14.0
+
+google-cloud-storage==2.5.0
+# pydrive2==1.14.0
 
 #datetime==
 #time==
